# PERF: CategoricalIndex.get_indexer