- [x] closes #42249
- [ ] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing.html#code-standards) for how to run them
- [ ] whatsnew entry

```
import pandas as pd
import string, itertools
data_unique = pd.CategoricalIndex(
            ["".join(perm) for perm in itertools.permutations(string.printable, 3)]
)
cat_list = ["a", "c"]

%timeit data_unique.get_indexer(cat_list)
106 µs ± 6.14 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
39 ms ± 616 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- 1.2.4
290 ms ± 3.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
```