@jreback 2021-06-28
@meeseeksdev backport 1.3.x