@jreback 2021-06-28
cc @jbrockmendel guess this won't backport easily